[externals] Update dependencies (#3664)
* zlib: 1.3.1.2 -> 1.3.2
* vulkan-validation-layers: 1.4.335.0 -> 1.4.341.0
* sirit: 1.0.3 -> 1.0.4
* httplib: 0.35.0 -> 0.37.0
* xbyak: 7.33.3 -> 7.35.2
* catch2: 3.12.0 -> 3.13.0
* vulkan-headers: 1.4.342 -> 1.4.345
* vulkan-utility-libraries: 1.4.342 -> 1.4.345

Also fixed a build error with newer xbyak.

Reviewed-on: https://git.eden-emu.dev/eden-emu/eden/pulls/3664

+++ src/dynarmic/src/dynarmic/backend/x64/emit_x64_saturation.cpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: Copyright 2025 Eden Emulator Project
+// SPDX-FileCopyrightText: Copyright 2026 Eden Emulator Project
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 /* This file is part of the dynarmic project.
@@ -135,7 +135,7 @@ void EmitX64::EmitSignedSaturation(EmitContext& ctx, IR::Inst* inst) {
 
     const u32 mask = (1u << N) - 1;
     const u32 positive_saturated_value = (1u << (N - 1)) - 1;
-    const u32 negative_saturated_value = 1u << (N - 1);
+    const u64 negative_saturated_value = 1u << (N - 1);
 
     const Xbyak::Reg32 result = ctx.reg_alloc.ScratchGpr(code).cvt32();
     const Xbyak::Reg32 reg_a = ctx.reg_alloc.UseGpr(code, args[0]).cvt32();
